requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dateutil
python-dotenv
ghasedak_sms